add checks before deleting employee

+++ src/components/ManpowerProfileCard.vue
@@ -4,24 +4,28 @@ import { getFirestore } from 'firebase/firestore';
 import { doc, collection, getDocs, deleteDoc } from 'firebase/firestore';
 import AddManpowerPopUp from '@/components/AddManpowerPopUp.vue';
 import { getStorage, ref, getDownloadURL, deleteObject } from 'firebase/storage';
+import { ref as vueref } from 'vue';
+import PopUp from '@/components/PopUp.vue';
 
 const storage = getStorage(app);
 
 export default {
   components: {
-    AddManpowerPopUp
-  },
-  data() {
-    return {
-      show: false
-    };
-  },
-  methods: {
-    showAddManpowerPopUp() {
-      this.show = true;
-    }
+    AddManpowerPopUp,
+    PopUp
   },
   setup() {
+    const show = vueref(false);
+    const showError = vueref(false);
+    const errorMessage = vueref('');
+    function showAddManpowerPopUp() {
+      show.value = true;
+    }
+
+    function showErrorPopUp() {
+      showError.value = true;
+    }
+
     const db = getFirestore(app);
 
     async function display() {
@@ -77,18 +81,55 @@ export default {
     }
     display();
 
-    async function deleteEmployee(employeeId, employeeName) {
-      alert('You are going to delete employee: ' + employeeName);
-      const docRef = doc(db, 'employees', employeeId); // use the document ID to create the document reference
-      await deleteDoc(docRef);
-      const imgRef = ref(storage, 'employee-' + employeeName + '.png');
-      // Delete the file
-      await deleteObject(imgRef);
-      console.log('Document successfully deleted!', employeeId);
+    const appts = vueref([]);
+    const getAppts = async (name) => {
+      appts.value = [];
+      console.log('getAppts Called');
+      const querySnapshot = await getDocs(collection(db, 'new-appointments'));
+      const slotArray = ['s1', 's2', 's3', 's4'];
+      const promises = querySnapshot.docs.map(async (docDates) => {
+        for (let j = 0; j < slotArray.length; j++) {
+          const querySnapshot1 = await getDocs(
+            collection(db, 'new-appointments/' + docDates.id + '/' + slotArray[j])
+          );
+          await Promise.all(
+            querySnapshot1.docs.map(async (docc) => {
+              let documentData = docc.data();
+              let apptDate = documentData.appt_date;
 
-      // remove the row from the HTML table
-      let row = document.querySelector(`[data-employee-id="${employeeId}"]`).closest('tr');
-      row.remove();
+              let groomer = documentData.appt_groomer;
+              if (name === groomer) {
+                appts.value.push(apptDate);
+                console.log(apptDate, ' pushed');
+              }
+            })
+          );
+        }
+      });
+      await Promise.all(promises);
+    };
+
+    async function deleteEmployee(employeeId, employeeName) {
+      await getAppts(employeeName);
+      console.log(appts.value);
+
+      if (appts.value.length > 0) {
+        errorMessage.value =
+          'Unable to delete. ' + employeeName + ' has pending appointments: ' + appts.value;
+        showErrorPopUp();
+      } else {
+        alert('You are going to delete employee: ' + employeeName);
+        const docRef = doc(db, 'employees', employeeId); // use the document ID to create the document reference
+        await deleteDoc(docRef);
+        const imgRef = ref(storage, 'employee-' + employeeName + '.png');
+        // Delete the file
+        await deleteObject(imgRef);
+        console.log('Document successfully deleted!', employeeId);
+
+        // remove the row from the HTML table
+        let row = document.querySelector(`[data-employee-id="${employeeId}"]`).closest('tr');
+        row.remove();
+      }
     }
 
     async function refresh() {
@@ -98,7 +139,15 @@ export default {
       }
       display();
     }
-    return { display, refresh };
+    return {
+      display,
+      refresh,
+      show,
+      showError,
+      errorMessage,
+      showAddManpowerPopUp,
+      showErrorPopUp
+    };
   }
 };
 </script>
@@ -111,6 +160,9 @@ export default {
   <div id="manpower-profile-cards">
     <table id="manpower-table" class="auto-index"></table>
   </div>
+  <PopUp id="error-popup" v-model="showError">
+    <h3>{{ errorMessage }}</h3>
+  </PopUp>
 </template>
 
 <style>
